fix(telegram): lazy-load sticker vision model lookup
