+++ extensions/telegram/src/sticker-vision.runtime.ts
@@ -1,0 +1,23 @@
+import {
+  findModelInCatalog,
+  loadModelCatalog,
+  modelSupportsVision,
+  resolveDefaultModelForAgent,
+} from "openclaw/plugin-sdk/agent-runtime";
+import type { OpenClawConfig } from "openclaw/plugin-sdk/config-runtime";
+
+export async function resolveStickerVisionSupportRuntime(params: {
+  cfg: OpenClawConfig;
+  agentId?: string;
+}): Promise<boolean> {
+  const catalog = await loadModelCatalog({ config: params.cfg });
+  const defaultModel = resolveDefaultModelForAgent({
+    cfg: params.cfg,
+    agentId: params.agentId,
+  });
+  const entry = findModelInCatalog(catalog, defaultModel.provider, defaultModel.model);
+  if (!entry) {
+    return false;
+  }
+  return modelSupportsVision(entry);
+}
